+++ packages/bot-detector/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tinyhttp/bot-detector
 
+## 0.1.16
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @tinyhttp/app@0.3.16
+
 ## 0.1.15
 
 ### Patch Changes