+++ packages/markdown/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tinyhttp/markdown
 
+## 0.2.16
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @tinyhttp/app@0.3.16
+
 ## 0.2.15
 
 ### Patch Changes