+++ packages/res/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tinyhttp/res
 
+## 0.1.13
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @tinyhttp/req@0.1.9
+
 ## 0.1.12
 
 ### Patch Changes